chore: move tests to lib.rs
